requirements.txt
python-telegram-bot[job-queue]==21.10
aiohttp==3.11.11
logging
